refactor(modularization): Phase 5 step 5-S4f — chat_collection_service (task #26)
Final move in the 5-S4 sequence: ``chat_collection_service`` moves
into ``aperag/domains/conversation/service/`` with clean
cross-domain imports of the Phase 3 and Phase 4 providers. Because
``#1629`` (KB) and ``#1633`` (identity / governance / marketplace /
model_platform) are both on main, the G1 legal cross-domain paths
finally exist:

* ``from aperag.db.models import Collection, User`` → split into
  ``from aperag.domains.knowledge_base.db.models import Collection``
  (direct cross-domain, G1 allows domain→domain). ``User`` is *not*
  imported: G16 canonical (``msg=6d2ae86a``) forbids User ORM
  imports outside the identity domain, so the ``_mark_as_chat_collection``
  closure issues a ``sqlalchemy.text("UPDATE users SET
  chat_collection_id = :cid WHERE id = :uid")`` call against the
  session instead of ``session.get(User, ...)`` + attribute
  assignment. The transaction semantics are identical; only the
  type-system coupling changes.
* ``from aperag.schema.view_models import CollectionConfig, ModelSpec,
  TagFilterCondition, TagFilterRequest`` → ``aperag.schema.common``
  for the two shared primitives + ``aperag.domains.model_platform.schemas``
  for the Phase 4-owned tag-filter shapes.
* ``from aperag.service.collection_service import collection_service``
  → ``aperag.domains.knowledge_base.service.collection_service``
  (Phase 3 direct cross-domain).
* ``from aperag.service.llm_available_model_service import
  llm_available_model_service`` →
  ``aperag.domains.model_platform.service.llm_available_model_service``
  (Phase 4 direct cross-domain).
* ``KnowledgeBaseCollectionView`` return type kept from Phase 3 Step
  5c — no benefit in replacing it with the concrete ``Collection``
  class here and it keeps the pre-move signature stable.
* ``"DELETED"`` / ``"CHAT"`` literal compares keep the Flag 1 / G15
  pattern — no enum import needed for ``CollectionStatus`` /
  ``CollectionType``.

``aperag/service/chat_collection_service.py`` collapses to a 32-line
re-export shim covering ``ChatCollectionService`` +
``chat_collection_service`` for the pre-migration caller base
(``views/chat.py`` + the ``_ChatInitOpsAdapter`` in ``app.py`` +
``aperag.service.chat_document_service`` DI wire-up).

Verify:
* ``from aperag.service.chat_collection_service import
  ChatCollectionService, chat_collection_service`` OK; class
  identity preserved against the domain home.
* ``pytest tests/unit_test/`` 673 passed / 29 skipped / 1 deselected
  (performance flake pre-existing) — importantly the G16 User-ORM
  ban (``test_phase4_consumer_domains_never_import_user_orm_class``)
  stays green because conversation does not import User.
* ``uvx ruff check`` + ``uvx ruff format --check`` clean.
* G1 on ``aperag/domains/conversation/service/chat_collection_service.py``
  zero legacy aggregate imports.
